% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.data.r
\docType{data}
\name{camera.data}
\alias{camera.data}
\title{Camera trap data for Diel.Niche}
\format{
A data frame with 2441 rows and 6 variables:
\describe{
  \item{species}{The common name of the species data.}
  \item{surveyID}{The site and season code for the respective opossum detection. The site is denoted by the first 8 characters and the season is the last four. WI12 = winter 2012, SP12 = spring 2012, SU12 = summer 2012, and FA12 = fall 2012.}
  \item{datetime}{The date & time (m-d-y h:m:s) of the respective opossum detection. Seconds are present in this column but were not actually included.}
  \item{timezone}{The timezone of the data as recognized by R (i.e., US/Central).}
  \item{lat}{The average latitude of this study area in decimal degrees (crs = 4326).}
  \item{lon}{The average longitude of this study area in decimal degrees (crs = 4326).}
}
}
\source{
\url{https://doi.org/10.1111/1365-2656.12967}
}
\usage{
camera.data
}
\description{
2012 camera trap data of Virginia opossum (Didelphis virginiana) from urban greenspace throughout Chicago, Illinois, USA
}
\keyword{datasets}
