% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.post.prob.r
\name{construct.post.prob}
\alias{construct.post.prob}
\title{Sets inputs for multinomineq::postprob}
\usage{
construct.post.prob(bf, hyp.set, prior, indicator)
}
\arguments{
\item{bf}{bayes factor outputs}

\item{hyp.set}{hypothesis set}

\item{prior}{Prior used for bayes factors. NULL indicates equal weights.}

\item{indicator}{identifies which algorithm/function was used to fit the data}
}
\value{
Internal list
}
\description{
A function to setup the inputs needed to estimate model posterior probabilities
}
\keyword{internal}
