% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.bf.r
\name{diel.bf}
\alias{diel.bf}
\title{Calculate Bayes factor (internal function)}
\usage{
diel.bf(
  y,
  idx.mod,
  diel.setup,
  reps,
  prior,
  n.mcmc,
  burnin,
  n.cpu,
  alt.optim,
  prints
)
}
\arguments{
\item{y}{vector of frequencies}

\item{idx.mod}{vector of indices indicating which hypotheses to use from diel.setup}

\item{diel.setup}{provided by user or used default as diel.setup=diel.ineq()}

\item{reps}{The number of replicate sets of three frequencies}

\item{prior}{Prior used for bayes factors. NULL indicates equal weights.}

\item{n.mcmc}{The number of mcmc iterations}

\item{burnin}{Burnin for mcmc algorithim}

\item{n.cpu}{Currently fixed at 1}

\item{alt.optim}{Alternative optimization for bayes factors}

\item{prints}{Whether to print information}
}
\value{
Internal list
}
\description{
Wrapper for multinomineq::bf_multinom
}
\keyword{internal}
