% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.hyp.func.r
\name{diel.hypotheses.func}
\alias{diel.hypotheses.func}
\title{Diel Modeling}
\usage{
diel.hypotheses.func(
  y,
  hyp.set,
  diel.setup = diel.ineq(),
  prior = NULL,
  n.mcmc = 50000,
  burnin = 10000,
  prints = TRUE,
  alt.optim = FALSE
)
}
\arguments{
\item{y}{Vector of frequencies in order of twilight, day, night.}

\item{hyp.set}{Vector of diel hypotheses names representing hypotheses set or individual hypotheses.}

\item{diel.setup}{A list of multinomial inequalities (Matrix A and vector b), representing diel hypotheses setup using the function 'diel.ineq'. Defaults to the defaults of diel.ineq function.}

\item{prior}{Prior probabilities for models. Defaults to equal among models.}

\item{n.mcmc}{Number of mcmc iterations.}

\item{burnin}{Burn-in number of mcmc iterations.}

\item{prints}{Whether to print messages about model fitting.}

\item{alt.optim}{Defaults to false. If TRUE, uses a different approach to derive the bayes factors. It can be more stable, but takes a bit longer.}
}
\value{
A list of outputs, including bayes factors for a model set, model bayes factor inputs, posterior samples, warning indicator, and posterior predictive checks.

A list of outputs
\item{bf.final}{Bayes factor for hyopthesis set}
\item{bf}{A list of ordered individual model bayes factor inputs}  
\item{posteriors}{A list of ordered matrices for model posterior distributions}  
\item{most.supported.model}{Character name of the most supported model detrerminded by the maximum probability of support from the bayes factors}  
\item{indicator.highest.bf.model}{An indicator of which ordered model is the most supported model}  
\item{indicator}{A vector indicating which models had errors and results were not returned. A 1 indicates a model was not fit and 0 indicates no errors.}  
\item{ppc}{A list of ordered model posterior predictive check output}
\item{ppc.ms}{Posterior predictive check output from the most supported model}    
\item{model.ms}{Posterior distributions of the most supported model}
}
\description{
Diel model hypotheses evaluation and parameter estimation.
This is essentially a wrapper function for functions provided by the package ‘multinomineq’.
}
\examples{
diel.hypotheses.func(y=y,hyp.set=hyp.set)
                    
Required libraries:   multinomineq, retry, MASS
}
