% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diel.Inequalities.r
\name{diel.ineq}
\alias{diel.ineq}
\title{Inequality Setup}
\usage{
diel.ineq(
  xi = NULL,
  e = NULL,
  e.D = NULL,
  e.N = NULL,
  e.CR = NULL,
  e.EC = NULL,
  e.AV = NULL,
  xi.D = NULL,
  xi.N = NULL,
  xi.CR = NULL,
  xi.C = NULL,
  xi.EC = NULL,
  p.avail = NULL,
  separation = NULL
)
}
\arguments{
\item{xi}{Default c(0.8, 0.1). The first element is the minimum probability of a singular hypothesis (e.g., Diurnal). The second element is the minimum probability for Cathemeral Traditional or Cathemeral General, or binomial hypotheses.}

\item{e}{Default is 0.05. A single value of variation for probabilities. If specified, it will be applied to all hypotheses, regardless of whether individual epsilon hypotheses values are specified.}

\item{e.D}{Default is 0.05. A single value of variation for the Diurnal hypothesis.}

\item{e.N}{Default is 0.05. A single value of variation for the Nocturnal hypothesis.}

\item{e.CR}{Default is 0.05. A single value of variation for the Crepuscular hypothesis.}

\item{e.EC}{Default is 0.05. A single value of variation for the Evan Cathemeral hypothesis.}

\item{e.AV}{Default is 0.10. A single value of variation for the Available Cathemeral hypothesis.}

\item{xi.D}{Default c(0.80,0.90). A vector of the lower threshold value and the most likely value, respectively for the Diurnal hypothesis.}

\item{xi.N}{Default c(0.80,0.90). A vector of the lower threshold value and most likely value, respectively for the Nocturnal hypothesis.}

\item{xi.CR}{Default c(0.80,0.90). A vector of the lower threshold value and most likely value, respectively for the Crepuscular hypothesis.}

\item{xi.C}{Default c(0.2). Lower value threshold to be cathemeral}

\item{xi.EC}{Default c(0.33). A single value of the available amount of time in all three diel periods.}

\item{p.avail}{Default c(0.166666,0.4166667). A vector of the available time in the periods of crepuscular and diurnal. Nighttime availability is found by subtraction.}

\item{separation}{Default is 0. However, you can separate the hypotheses to create empty space between hypotheses probabilit space}
}
\value{
diel.hyp A list of diel hypotheses as multinomial inequalities.
\item{inputs}{Includes all inputted values; epsilon, xi, and p.avail.} 
\item{D.th, N.th, CR.th, EC.th, C.th, D.max, N.max, CR.max, D.var, N.var, CR.var, C.var, AC.var, AV.var,
   Uncon, C.max, EC, AV.EQ}{Each is a list of three elements: Hypotheis Descriptive Name, A matrix, and b vector.}
}
\description{
Multinomial model inequalities for diel hypotheses
}
\examples{
diel.ineq()
diel.ineq(e=0.01) #To replace all epsilon values with 0.01.
#To replace the default values with a new epsilon value and a new xi value.
}
