% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.plot.r
\name{diel.plot}
\alias{diel.plot}
\title{Plot Diel Hypotheis}
\usage{
diel.plot(hyp, diel.setup = NULL, posteriors = NULL)
}
\arguments{
\item{hyp, }{hypothesis code name to use}

\item{diel.setup}{Defaults to using diel.ineq function. A list of multinomial inequalities (Matrix A and vector b), representing diel hypotheses setup using the function 'diel.ineq'.}

\item{posteriors}{A single models MCMC output from the function 'diel.hypotheses.func'.}
}
\value{
A plotly 3d plot
}
\description{
Plots the diel niche space and posterior disribution of a fitted model.
}
\examples{
out=diel.fit(y=t(matrix(c(11,87,2))),hyp.set="D.max",n.mcmc=1000,burnin=200)
diel.plot(hyp="D.max",posteriors=out$post.samp[[1]])
}
