% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.post.r
\name{diel.post}
\alias{diel.post}
\title{Samples posterior distributions}
\usage{
diel.post(
  y,
  idx.mod,
  diel.setup,
  bf.Ab.new,
  reps,
  n.chains,
  n.mcmc,
  burnin,
  n.cpu,
  indicator,
  prints
)
}
\arguments{
\item{y, }{vector of frequencies}

\item{idx.mod}{vector of indices indicating which hypotheses to use from diel.setup}

\item{diel.setup}{provided by user or used default as 'diel.setup=diel.ineq()'}

\item{bf.Ab.new}{Modified A matrix and b vector for those equality/inequality hypotheses}

\item{reps}{The number of replicate sets of three frequencies}

\item{n.chains}{The number of chains for posterior sampling}

\item{n.mcmc}{The number of mcmc iterations}

\item{burnin}{Burnin for mcmc algorithim}

\item{n.cpu}{Currently fixed at 1}

\item{indicator}{Indicates failed bayes factor}

\item{prints}{Whether to print information}
}
\value{
Internal list
}
\description{
Wrapper for multinomineq::sampling_multinom
}
\keyword{internal}
