% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.prob.hyp.r
\name{find.prob.hyp}
\alias{find.prob.hyp}
\title{Inequality Setup}
\usage{
find.prob.hyp(hyp, diel.setup = NULL)
}
\arguments{
\item{hyp}{Hypothesis name: D.th, D.max, D.var, Dn.th, Dn.max, Dn.var, Dc.th, Dc.max, Dc.var, Dcr.th, Dcr.max,Dcr.var, N.th, N.max, N.var, Nd.th, Nd.max, Nd.var, Nc.th, Nc.max, Nc.var, Ncr.th, Ncr.max, Ncr.var, CR.th, CR.max, CR.var,CRd.th,' CRd.max, CRd.var, CRn.th, CRn.max, CRn.var, CRc.th, CRc.max, CRc.var,'EC.th, EC.var, AC.var.}

\item{diel.setup}{A list of multinomial inequalities (Matrix A and vector b), representing diel hypotheses setup using the function 'diel.ineq'.}
}
\value{
A matrix of probabilities that match hypothesis in variable hyp
}
\description{
Function that inputs a given hypothesis and outputs as many possible
probability sets that match the diel hypothesis (i.e., satisfies the inequality constrints).
}
\examples{
find.prob.hyp(hyp ="D.max")
}
