% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hyp.name.r
\name{full.hyp.name}
\alias{full.hyp.name}
\title{Kernel Overlap density integration}
\usage{
full.hyp.name(ms.model)
}
\arguments{
\item{ms.model}{most supported model/hypothesis}
}
\value{
Internal list
}
\description{
Gets full name of hypothesis
}
\keyword{internal}
