% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp.sets.r, R/hyp.sets2.r
\name{hyp.sets}
\alias{hyp.sets}
\title{Hypothesis Sets}
\usage{
hyp.sets(hyp.in = NULL)

hyp.sets(hyp.in = NULL)
}
\arguments{
\item{hyp.in, }{Hypothesis set code names}
}
\value{
Names of hypotheses for the set. If NULL, the names of all hypotheses sets are returned.

Names of hypotheses for the set. If NULL, the names of all hypotheses sets are returned.
}
\description{
Call defined hypotheses sets

Call defined hypotheses sets
}
\examples{
hyp.sets()
hyp.sets("hyp.th")
hyp.sets()
hyp.sets("hyp.th")
}
