% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp.sets2.r
\name{hyp.sets2}
\alias{hyp.sets2}
\title{Hypothesis Sets}
\usage{
hyp.sets2(hyp.in = NULL)
}
\arguments{
\item{hyp.in, }{Hypothesis set code names}
}
\value{
Names of hypotheses for the set. If NULL, the names of all hypotheses sets are returned.
}
\description{
Call defined hypotheses sets
}
\examples{
hyp.sets()
hyp.sets("hyp.th")
}
