% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.Ab.r
\name{modify.Ab}
\alias{modify.Ab}
\title{Check on inputs}
\usage{
modify.Ab(y, idx.mod, reps, diel.setup, delta)
}
\arguments{
\item{y}{vector of frequencies}

\item{idx.mod}{indices of hypotheses of diel.setup to use}

\item{reps}{replicate samples of three}

\item{diel.setup}{provided by user or used default as 'diel.setup=diel.ineq()'}

\item{delta}{error tolerance allowed}
}
\value{
Internal list
}
\description{
A function to check the inputs provided through the function 'diel.fit'
}
\keyword{internal}
