% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.setup.hyp.params.r
\name{plot.setup.hyp.params}
\alias{plot.setup.hyp.params}
\title{Setup plotting hypotheses (Internal)}
\usage{
\method{plot}{setup.hyp.params}(diel.setup, index.models)
}
\arguments{
\item{diel.setup}{provided by user or used default as diel.setup=diel.ineq()}

\item{index.models}{vector of indices indicating which hypotheses to use from diel.setup}
}
\value{
Internal list
}
\description{
Prepares data to plot hypotheses
}
\keyword{internal}
