% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.niche.r
\name{posthoc.niche}
\alias{posthoc.niche}
\title{Posthoc diel niche}
\usage{
posthoc.niche(y, hyp, diel.setup = NULL)
}
\arguments{
\item{y}{a matrix of probabilities of twilight, daytime, and nightime (three columns). Each row is a replicate.}

\item{hyp}{Vector of diel hypotheses names representing hypotheses set or individual hypotheses.}

\item{diel.setup}{A list of multinomial inequalities (Matrix A and vector b), representing diel hypotheses setup using the function 'diel.ineq'. If not provided, it will use the defaults of the diel.ineq function.}
}
\value{
data.frame of the probabilities of activity during twilight, daytime, nighttime, and matched hypothesis from the hypothesis set provided.
}
\description{
Diel model hypotheses based on posthoc probabilitiy values
For each input of twilight, daytime, and nightime probabilities, the function will return the defined diel hyppthesis
}
