% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.hyp.plot.params.r
\name{setup.hyp.plot.params}
\alias{setup.hyp.plot.params}
\title{Setup plotting hypotheses (Internal)}
\usage{
setup.hyp.plot.params(diel.setup, index.models, more.points = FALSE, ...)
}
\arguments{
\item{diel.setup}{provided by user or used default as diel.setup=diel.ineq()}

\item{index.models}{vector of indices indicating which hypotheses to use from diel.setup}

\item{more.points}{which to use, p.options or p.options2}

\item{...}{Other parameters}
}
\value{
Internal list
}
\description{
Prepares data to plot hypotheses
}
\keyword{internal}
