% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.diel.r
\name{sim.diel}
\alias{sim.diel}
\title{Simulate}
\usage{
sim.diel(
  n.sim = 1,
  reps = 1,
  n.sample = 100,
  hyp,
  diel.setup = NULL,
  sd.error = 0
)
}
\arguments{
\item{n.sim}{The number of simulated datasets}

\item{reps}{The number of sets of probabilities to use when simulating crepuscular, daytime, and nocturnal frequencies}

\item{n.sample}{The number of total samples for a given simulation}

\item{hyp}{The hypothesis to simulate data from}

\item{diel.setup}{Multinomial inequalities for hypotheses setup using function 'diel.ineq'.}

\item{sd.error}{Normal distribution standard deviation to simulate error to add to the probabilities on the logit-scale. Default is 0}
}
\value{
A list of outputs
\item{y}{Matrix of simulated datasets}
\item{p}{Probabilities used to simulate the data}
}
\description{
Simulate diel data
}
\examples{
sim.diel(n.sim=1,reps=1,n.sample=100,hyp="D.th")
}
