% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.time.records.R
\name{trim.time}
\alias{trim.time}
\title{Thin detections based on a time threshold}
\usage{
trim.time(
  data,
  datetime.column,
  site.column,
  species.column = NULL,
  minutes.between = 15,
  progress.bar = TRUE
)
}
\arguments{
\item{data}{A `data.frame` containing detection data. Must include a datetime column,
a site identifier column, and optionally a species column.}

\item{datetime.column}{A character string giving the name of the column in `data` containing
datetime information. The datetime column must be of class `POSIXct` or `POSIXt`.}

\item{site.column}{A character string giving the name of the column in `data` identifying the site
where each detection occurred.}

\item{species.column}{Optional. A string giving the name of the column in `data`
identifying the species detected. If `NULL` (default), all records within a site are
assumed to be of the same species. Thus, if the dataset contains multiple species you
should use this argument.}

\item{minutes.between}{A positive number indicating the minimum number of minutes
that must pass between detections at a site (and species, if specified) for both to be retained.
Defaults to 15.}

\item{progress.bar}{Logical; if `TRUE`, displays a progress bar during processing.
Defaults to `TRUE`.}
}
\value{
A `data.frame` containing the same columns as the input `data`, but with records
  removed so that no two retained records within the same site (and species, if specified)
  are within `minutes.between` minutes of each other.
}
\description{
Filters detection records so that observations within a specified number of minutes
from the previous observation (at the same site) are removed. For multi-species
datasets, records should be thinned by site and species detected at that site.
This function can be used to thin, for example, camera trap or acoustic data where multiple
detections in rapid succession may not be independent.
}
\details{
This function first splits the data by site (and species, if provided), then iteratively removes
records that occur within `minutes.between` minutes of a previous detection. It handles
sequences of closely spaced detections intelligently to retain the earliest and ensure spacing.
}
\examples{
\dontrun{
# Example with datetime in POSIXct format
dat <- data.frame(
  site = rep("A", 5),
  species = rep("red fox", 5),
  datetime = as.POSIXct(c(
    "2023-01-01 00:00:00",
    "2023-01-01 00:05:00",
    "2023-01-01 00:20:00",
    "2023-01-01 00:21:00",
    "2023-01-01 00:40:00"
  ))
)
# Species column not needed as this dataset
#  has one species
response <- trim.time(
  dat,
  datetime.column = "datetime",
  site.column = "site",
  minutes.between = 15
)
          
  dat <- data.frame(
  site = rep("A", 10),
  species = rep(c("red fox", "coyote"), each = 5),
  datetime = as.POSIXct(c(
    "2023-01-01 00:00:00",
    "2023-01-01 00:05:00",
    "2023-01-01 00:20:00",
    "2023-01-01 00:21:00",
    "2023-01-01 00:40:00",
    "2023-01-01 00:00:00",
    "2023-01-01 00:05:00",
    "2023-01-01 00:20:00",
    "2023-01-01 00:21:00",
    "2023-01-01 00:40:00"
  ))
)
# Species solumn needed as this dataset
#  has more than one species.
response <- trim.time(
  dat,
  datetime.column = "datetime",
  species.column = "species",
  site.column = "site",
  minutes.between = 15
)

}
}
