% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.hyp.r
\name{what.hyp}
\alias{what.hyp}
\title{Hypothesis Codes}
\usage{
what.hyp(hyp.in = NULL)
}
\arguments{
\item{hyp.in}{hypothesis code name, NULL, or ?}
}
\value{
Full name of hypothesis if code name is provided. If NULL a general description of hypotheses is provided. If ? then hypotheses code names are provided.
}
\description{
Call defined hypotheses sets
}
\examples{
what.hyp()
what.hyp("D.th")
what.hyp("?")
}
