% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.diel.times.R
\name{bin.diel.times}
\alias{bin.diel.times}
\title{Bin Datetime Values into Diel Categories}
\usage{
bin.diel.times(
  data,
  datetime.column,
  lat.column,
  lon.column,
  bin.type.list = make.diel.bin.list(plot.bins = FALSE),
  na_vals = c("remove", "error")
)
}
\arguments{
\item{data}{A data frame containing observation data, including datetime, latitude, and longitude columns, and an optional species column if your
dataset has multiple species.}

\item{datetime.column}{Character string giving the name of the POSIXct datetime column.}

\item{lat.column}{Character string giving the name of the latitude column (in decimal degrees).}

\item{lon.column}{Character string giving the name of the longitude column (in decimal degrees).}

\item{bin.type.list}{A list of logical expressions generated with [make.diel.bin.list()]. Must be of class `diel.bin.list`.}

\item{na_vals}{Character; how to handle NA values returned by `suncalc::getSunlightTimes()`. Must be either `"remove"` (default) or `"error"`. See
details for more information.}
}
\value{
A copy of the input data frame with an additional `dielBin` column specifying the assigned diel bin for each observation.
}
\description{
Assigns observations to diel bins (e.g., day, night, dawn, dusk, twilight) based on
sun angle transitions calculated with `suncalc::getSunlightTimes()`. The function uses
user-defined logical expressions stored in a `diel.bin.list` to classify times.
}
\details{
The function calculates solar transition times (e.g., sunrise, sunset, dawn, dusk) using the `suncalc` package. It then evaluates
user-specified logical expressions to determine which bin each observation falls into. The bin definitions must be specified
using a \code{diel.bin.list} object created with [make.diel.bin.list()]. This allows users to either define how each bin is
constructed with logical expressions based on solor transitions, or to use the default bins provided.



# Example Expression Format
}
\examples{
# Example with default bin.list
data(camera.data)

# remove images of the same species that occurred
#  within 15 minutes of a previous photo record
#  by site.
thinned_data <- trim.time(
  data = camera.data,
  datetime.column = "datetime",
  site.column = "surveyID",
  species.column = "species",
  progress.bar = FALSE
)

# construct bin list, using defaults
bin_list <- make.diel.bin.list(
  plot.bins = FALSE
)

# bin diel times
thinned_data <- bin.diel.times(
  data = thinned_data,
  datetime.column = "datetime",
  lat.column = "lat",
  lon.column = "lon",
  bin.type.list = bin_list
)

# calculate frequency of each bin,
#  this could then be
y <- table(
  thinned_data$species,
  thinned_data$dielBin
)
}
\seealso{
[make.diel.bin.list()]
}
\author{
Mason Fidino
}
