% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.inputs.r
\name{check.inputs}
\alias{check.inputs}
\title{Check on inputs}
\usage{
check.inputs(y, hyp.set, prior, bf.fit, diel.setup)
}
\arguments{
\item{y}{vector of frequencies}

\item{hyp.set}{the hypotheses to use}

\item{prior}{Prior used for bayes factors. NULL indicates equal weights.}

\item{bf.fit}{Indicator whether to fit bayes factors}

\item{diel.setup}{provided by user or used default as 'diel.setup=diel.ineq()'}
}
\value{
Internal list
}
\description{
A function to check the inputs provided through the function 'diel.fit'
}
\keyword{internal}
