% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.fit.r
\name{diel.fit}
\alias{diel.fit}
\title{Diel Modeling}
\usage{
diel.fit(
  y,
  hyp.set,
  bf.fit = TRUE,
  diel.setup = NULL,
  prior = NULL,
  n.mcmc = 50000,
  burnin = 10000,
  prints = TRUE,
  alt.optim = FALSE
)
}
\arguments{
\item{y}{a matrix of frequencies of animal detections. Each row is a replicate dataset. Rows should be limited when using P-s hyps (1 or 2). The matrix should always be three columns in this order: twilight, day, night.}

\item{hyp.set}{Vector of diel hypotheses names representing hypotheses set or individual hypotheses.}

\item{bf.fit}{If TRUE, will calculate bayes factors for the model sit. Default is TRUE.}

\item{diel.setup}{A list of multinomial inequalities (Matrix A and vector b), representing diel hypotheses setup using the function 'diel.ineq'. If not provided, it will use the defaults of the diel.ineq function.}

\item{prior}{Prior probabilities for models used in bayes factors. Defaults to equal among models.}

\item{n.mcmc}{Number of mcmc iterations.}

\item{burnin}{Burn-in number of mcmc iterations.}

\item{prints}{Whether to print messages about model fitting.}

\item{alt.optim}{Default is FALSE. If TRUE, uses an alternative approach to derive the bayes factors. It can be more stable, but takes a bit longer.}
}
\value{
A list of outputs, including bayes factors for a model set, model bayes factor inputs, posterior samples, warning indicator, and posterior predictive checks.

A list of outputs
\item{bf.table}{Bayes factor for hyopthesis set}
\item{bf}{A list of ordered individual model bayes factor inputs}  
\item{post.samp}{A list of ordered matrices for model posterior distributions}  
\item{ms.model}{The name of the most supported model detrerminded by the maximum probability of support from the bayes factors}  
\item{ppc}{A list of ordered model posterior predictive check output}
\item{ms.ppc}{Posterior predictive check output from the most supported model}    
\item{post.samp.ms.model}{Posterior distributions of the most supported model}
}
\description{
Diel model hypotheses evaluation and parameter estimation.
This is essentially a wrapper function for functions provided by the package multinomineq.
}
\examples{
diel.hypotheses.func(y=t(matrix(c(10,100,10))),hyp.set=hyp.sets("hyp.th"))
                    
Required libraries:   multinomineq, retry, MASS
}
