% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diel.Inequalities.r
\name{diel.ineq}
\alias{diel.ineq}
\title{Inequality Setup}
\usage{
diel.ineq(
  xi = NULL,
  e = NULL,
  e.D = NULL,
  e.N = NULL,
  e.CR = NULL,
  e.EC = NULL,
  e.AV = NULL,
  xi.t.D = NULL,
  xi.t.N = NULL,
  xi.t.CR = NULL,
  xi.t.C = NULL,
  eta.D = NULL,
  eta.N = NULL,
  eta.CR = NULL,
  eta.C = NULL,
  p.avail = NULL,
  separation = NULL
)
}
\arguments{
\item{xi}{Default c(0.8, 0.1). The first element is the minimum threshold probability of singular hypotheses (e.g., Diurnal; Traditional Hypothesis set). The second element is the minimum probability for the General Hypothesis set. See details for additional information.}

\item{e}{Default is 0.10. A single value of variation for probabilities. If specified, it will be applied to all hypotheses, regardless of whether individual epsilon hypotheses values are specified.}

\item{e.D}{Default is 0.10. A single value of variation for the Diurnal hypothesis (Variation Hypothesis Set).}

\item{e.N}{Default is 0.10. A single value of variation for the Nocturnal hypothesis (Variation Hypothesis Set).}

\item{e.CR}{Default is 0.10. A single value of variation for the Crepuscular hypothesis (Variation Hypothesis Set).}

\item{e.EC}{Default is 0.10. A single value of variation for the Evan Cathemeral hypothesis (Variation Hypothesis Set).}

\item{e.AV}{Default is 0.10. A single value of variation for the Available Cathemeral hypothesis.}

\item{xi.t.D}{Default c(0.80). A single value of the lower threshold value for the Diurnal hypothesis (Threshold Hypothesis Set).}

\item{xi.t.N}{Default c(0.80). A single value of the lower threshold value for the Nocturnal hypothesis (Threshold Hypothesis Set)}

\item{xi.t.CR}{Default c(0.80). A single value of the lower threshold value for the Crepuscular hypothesis (Threshold Hypothesis Set)}

\item{xi.t.C}{Default c(0.2). A single value of the lower threshold value for the Cathemeral hypothesis (Threshold Hypothesis Set)}

\item{eta.D}{Default c(0.90). A single value of the most probable value for the Diurnal hypothesis (Variation Hypothesis Set)}

\item{eta.N}{Default c(0.90). A single value of the most probable value for the Nocturnal hypothesis (Variation Hypothesis Set)}

\item{eta.CR}{Default c(0.90). A single value of the most probable value for the Crepuscular hypothesis (Variation Hypothesis Set)}

\item{eta.C}{Default c(0.33). A single value of the most probable value for the Cathemeral hypothesis (Variation Hypothesis Set)}

\item{p.avail}{Default c(0.166666,0.4166667). A vector of the available time in the periods of crepuscular and diurnal. Nighttime availability is found by subtraction.}

\item{separation}{Default is 0. However, you can separate the hypotheses to create empty space between hypotheses probability space}
}
\value{
diel.hyp A list of diel hypotheses as multinomial inequalities.
\item{inputs}{Includes all inputted values; epsilon, xi, and p.avail.} 
\item{D.th, N.th, CR.th, EC.th, C.th, D.max, N.max, CR.max, D.var, N.var, CR.var, C.var, AC.var, AV.var,
   Uncon, C.max, EC, EQ.avail, D.avail,TW.avail,N.avail, D.TW.avail, N.TW.avail, D.N.avail}{Each is a list of three elements: Hypotheis Descriptive Name, A matrix, and b vector.}
}
\description{
Multinomial model inequalities for diel hypotheses. Given a numeric set of values between 0 and 1, \code{diel.ineq()}
will generate the inequality constraints that can be used within \code{\link{diel.fit}}.
}
\details{
In the event that xi is a scalar, the second value will be calculated as \code{(1-xi)/2},
 where \code{xi} is the scalar provided by the user. 
 
 The values provided here generated the requisite matrix \eqn{\boldsymbol{A}}
 and vector \eqn{\boldsymbol{b}}. For additional details on how these constraints
 are used in a multinomial model, see Heck and Davis-Stober (2019).
}
\examples{
diel.ineq()
diel.ineq(e=0.01) #To replace all epsilon values with 0.01.
#To replace the default values with a new epsilon value and a new xi value.
}
\references{
Heck, D. W., & Davis-Stober, C. P. (2019). Multinomial models with linear inequality
 constraints: Overview and improvements of computational methods for Bayesian
 inference. Journal of mathematical psychology, 91, 70-87.
}
