% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.plot.r
\name{diel.plot}
\alias{diel.plot}
\title{Plot Diel Hypothesis or Hypothesis Set Along with Posterior Samples}
\usage{
diel.plot(
  hyp,
  diel.setup = NULL,
  posteriors = NULL,
  more.points = FALSE,
  x.scene = 2.5,
  y.scene = 1,
  z.scene = 0.3,
  axis.size = 16,
  axis.lab.size = 18,
  legend.lab.size = 15
)
}
\arguments{
\item{hyp}{a vector of hypothesis code names (characters)}

\item{diel.setup}{If NULL uses the default diel.ineq function. Otherwise, a list of multinomial inequalities (Matrix A and vector b) representing diel hypotheses and the function needed for model fitting.}

\item{posteriors}{A single models MCMC output from the function 'diel.fit'.}

\item{more.points}{To use more points for hyps in plotting. Default is FALSE.}

\item{x.scene}{3d graphical parameter}

\item{y.scene}{3d graphical parameter}

\item{z.scene}{3d graphical parameter}

\item{axis.size}{3d graphical parameter}

\item{axis.lab.size}{3d graphical parameter}

\item{legend.lab.size}{3d graphical parameter}
}
\value{
A plotly 3d plot
}
\description{
Plots the diel niche space and posterior disribution of a fitted model.
}
\examples{
out=diel.fit(y=t(matrix(c(11,87,2))),hyp="D",post.fit=TRUE)
diel.plot(hyp="D",posteriors=out$post.samp.ms.model)
}
