% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.plot.r
\name{diel.plot}
\alias{diel.plot}
\title{Plot Diel Hypothesis or Hypothesis Set Along with Posterior Samples}
\usage{
diel.plot(
  fit = NULL,
  hyp = NULL,
  diel.setup = NULL,
  posteriors = NULL,
  more.points = FALSE,
  x.scene = 2.5,
  y.scene = 1,
  z.scene = 0.3,
  axis.size = 16,
  axis.lab.size = 18,
  legend.lab.size = 15
)
}
\arguments{
\item{fit}{a list object output from the function 'diel.fit'.}

\item{hyp}{a vector of hypothesis code names (characters)}

\item{diel.setup}{If NULL uses the default diel.ineq function. If a fit object is provided it will come from this object. Otherwise, a list of multinomial inequalities (Matrix A and vector b) representing diel hypotheses and the function needed for model fitting.}

\item{posteriors}{Posterior samples output from the function 'diel.fit'.}

\item{more.points}{To use more points for hyps in plotting. Default is FALSE.}

\item{x.scene}{3d graphical parameter}

\item{y.scene}{3d graphical parameter}

\item{z.scene}{3d graphical parameter}

\item{axis.size}{3d graphical parameter}

\item{axis.lab.size}{3d graphical parameter}

\item{legend.lab.size}{3d graphical parameter}
}
\value{
A plotly 3d plot
}
\description{
Plots the diel niche space and posterior distribution of a fitted model.
}
\details{
This function can be a little more difficult to get running for users who 
program in RStudio. Currently, a user would need to go to Tools > Global Options >
General > Advanced, change the rendering engine to Desktop OpenGL, and then restart RStudio.
}
\examples{
out=diel.fit(y=cbind(11,87,2),hyp="D",post.fit=TRUE)
diel.plot(out)
}
