% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel.post.r
\name{diel.post}
\alias{diel.post}
\title{Samples posterior distributions}
\usage{
diel.post(
  y,
  idx.mod,
  diel.setup,
  reps,
  n.mcmc,
  burnin,
  n.cpu,
  indicator,
  prints
)
}
\arguments{
\item{y, }{vector of frequencies}

\item{idx.mod}{vector of indices indicating which hypotheses to use from diel.setup}

\item{diel.setup}{provided by user or used default as diel.setup=diel.ineq()}

\item{reps}{The number of replicate sets of three frequencies}

\item{n.mcmc}{The number of mcmc iterations}

\item{burnin}{Burnin for mcmc algorithim}

\item{n.cpu}{Currently fixed at 1}

\item{prints}{Whether to print information}

\item{prior}{Prior used for bayes factors. NULL indicates equal weights.}

\item{alt.optim}{Alternative optimization for bayes factors}
}
\value{
Internal list
}
\description{
Wrapper for multinomineq::sampling_multinom
}
\keyword{internal}
