% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp.sets.r
\name{hyp.sets}
\alias{hyp.sets}
\title{Hypothesis Sets}
\usage{
hyp.sets(hyp.in)
}
\arguments{
\item{hyp.in, }{hypothesis code name to use}
}
\value{
names of hypotheses for that set
}
\description{
Call defined hypotheses sets
}
\examples{
hyp.sets("hyp.th")
}
