% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.Overlap.r
\name{prob.Overlap}
\alias{prob.Overlap}
\title{Kernel Overlap density integration}
\usage{
prob.Overlap(densityplot, dawn = c(6, 7), dusk = c(17, 18))
}
\arguments{
\item{densityplot}{a densityPlot object from package Overlap.}

\item{dawn}{begining and end numeric (0-24) times for dawn}

\item{dusk}{begining and end numeric (0-24) times for dusk}
}
\value{
Internal list
}
\description{
Integrate kernel density to derieve probability of twilight, daytime, nightime
}
\keyword{internal}
