% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.diel.r
\name{sim.diel}
\alias{sim.diel}
\title{Simulate}
\usage{
sim.diel(n.sim = 1, reps = 1, n.sample = 100, hyp, diel.setup = NULL)
}
\arguments{
\item{n.sim}{The number of simulated datasets}

\item{reps}{The number of replicates of crepuscular, daytime, and nocturnal for each simulated dataset}

\item{n.sample}{The number of total samples for a given simulation}

\item{hyp}{The hypothesis to simulate data from}

\item{diel.setup}{Multinomial inequalities for hypotheses setup using function 'diel.ineq'.}
}
\value{
A list of outputs
\item{y}{Matrix of simulated datasets}
\item{p}{Probabilities used to simulate the data}
}
\description{
Simulate diel data
}
\examples{
sim.diel(n.sim=1,reps=1,n.sample=100,hyp="D.th")
                    
Required libraries:   stats
}
